refactor to DDD
